receipt-non-agent/requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx
anthropic
gspread
google-auth
python-dotenv
openpyxl
